# docs/swagger/swagger.yaml
basePath: /api
definitions:
  gin-fast_app_models.DeleteRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  gin-fast_app_models.UpdateRequest:
    properties:
      deptId:
        type: integer
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      nickName:
        type: string
      phone:
        type: string
      roles:
        items:
          type: integer
        type: array
      sex:
        type: string
      status:
        type: integer
      userName:
        type: string
    required:
    - deptId
    - id
    - nickName
    - roles
    - sex
    - userName
    type: object
  models.AddRequest:
    properties:
      deptId:
        type: integer
      description:
        type: string
      email:
        type: string
      nickName:
        type: string
      password:
        type: string
      phone:
        type: string
      roles:
        items:
          type: integer
        type: array
      sex:
        type: string
      status:
        type: integer
      userName:
        type: string
    required:
    - deptId
    - nickName
    - roles
    - sex
    - userName
    type: object
...
  models.UpdateAccountRequest:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      phone:
        type: string
    required:
    - id
    - phone
    type: object
...
host: localhost:8080
info:
  contact:
    email: your-email@example.com
    name: API Support
    url: https://github.com/your-repo/gin-fast/issues
  description: 基于Gin框架的快速开发脚手架API文档
  license:
    name: MIT
    url: https://github.com/your-repo/gin-fast/blob/master/LICENSE
  termsOfService: https://github.com/your-repo/gin-fast
  title: Gin-Fast API
  version: "1.0"
paths:
...
  /users/{id}:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 用户ID格式错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 根据ID获取用户
      tags:
      - 用户管理
  /users/add:
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.AddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 用户创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 新增用户
      tags:
      - 用户管理
  /users/delete:
    delete:
      consumes:
      - application/json
      description: 删除用户信息
      parameters:
      - description: 用户删除请求参数
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/gin-fast_app_models.DeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 用户删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户管理
  /users/edit:
    put:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/gin-fast_app_models.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 用户更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户资料
      tags:
      - 用户管理
  /users/list:
    get:
      consumes:
      - application/json
      description: 获取用户列表，支持分页和过滤
      parameters:
      - default: 1
        description: 页码
        in: query
        name: pageNum
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 用户名或昵称
        in: query
        name: name
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 用户列表
      tags:
      - 用户管理
...
  /users/profile:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的信息，包括角色和权限
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回用户信息
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 用户未登录
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前登录用户信息
      tags:
      - 用户管理
  /users/updateAccount:
    put:
      consumes:
      - application/json
      description: 更新用户密码、手机号及邮箱信息
      parameters:
      - description: 用户账户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 用户账户信息更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户账户信息
      tags:
      - 用户管理
  /users/uploadAvatar:
    post:
      consumes:
      - multipart/form-data
      description: 上传用户头像文件
      parameters:
      - description: 用户头像文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 用户头像上传成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 上传用户头像
      tags:
      - 用户管理
swagger: "2.0"